✨[#74] 찜하기 수 top 5 Redis - 주석추가
